+++ src/util_generic.cc
@@ -42,11 +42,11 @@ string ies(int number) {
     return(number == 1 ? "y" : "ies");
 }
 
-string plural(int number, string text) {
+string plural(size_t number, string text) {
     return (to_string(number) + " " + text + (number == 1 ? "" : "s"));
 }
 
-string plurali(int number, string text) {
+string plurali(size_t number, string text) {
     return (to_string(number) + " " + text + (number == 1 ? "y" : "ies"));
 }
